pyproject.toml
[tool.poetry]
name = "auto_dns"
version = "0.1.0"
description = "A DNS Python Library."
authors = ["Ali Tavallaie <a.tavallaie@gmail.com>"]
readme = "README.rst"

[tool.poetry.scripts]
autodns = "auto_dns:app"

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0.1"
requests = "^2.31.0"
typer = "^0.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
